destination-postgres: bump CDK
